Init API server
