+++ api/README.md
@@ -1,0 +1,25 @@
+# Go API Server for swagger
+
+Planet Pulse is an API designed to serve climate data pulled from NOAA's Global Monitoring Laboratory FTP server. This API is based on the OpenAPI v3 specification.
+
+## Overview
+This server was generated by the [swagger-codegen]
+(https://github.com/swagger-api/swagger-codegen) project.  
+By using the [OpenAPI-Spec](https://github.com/OAI/OpenAPI-Specification) from a remote server, you can easily generate a server stub.  
+-
+
+To see how to make this your own, look here:
+
+[README](https://github.com/swagger-api/swagger-codegen/blob/master/README.md)
+
+- API version: 0.1.0
+- Build date: 2021-07-21T22:57:10.672601-04:00[America/New_York]
+
+
+### Running the server
+To run the server, follow these simple steps:
+
+```
+go run main.go
+```
+
