--- api/server/go/logger.go
@@ -1,32 +1,0 @@
-/*
- * Planet Pulse
- *
- * Planet Pulse is an API designed to serve climate data pulled from NOAA's Global Monitoring Laboratory FTP server. This API is based on the OpenAPI v3 specification.
- *
- * API version: 0.1.0
- * Contact: planetpulse.api@gmail.com
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-package swagger
-
-import (
-    "log"
-    "net/http"
-    "time"
-)
-
-func Logger(inner http.Handler, name string) http.Handler {
-    return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-        start := time.Now()
-
-        inner.ServeHTTP(w, r)
-
-        log.Printf(
-            "%s %s %s %s",
-            r.Method,
-            r.RequestURI,
-            name,
-            time.Since(start),
-        )
-    })
-}
