--- api/server/go/model_error.go
@@ -1,17 +1,0 @@
-/*
- * Planet Pulse
- *
- * Planet Pulse is an API designed to serve climate data pulled from NOAA's Global Monitoring Laboratory FTP server. This API is based on the OpenAPI v3 specification.
- *
- * API version: 0.1.0
- * Contact: planetpulse.api@gmail.com
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-package swagger
-
-type ModelError struct {
-
-	Code int32 `json:"code"`
-
-	Message string `json:"message"`
-}
