--- api/server/go/routers.go
@@ -1,78 +1,0 @@
-/*
- * Planet Pulse
- *
- * Planet Pulse is an API designed to serve climate data pulled from NOAA's Global Monitoring Laboratory FTP server. This API is based on the OpenAPI v3 specification.
- *
- * API version: 0.1.0
- * Contact: planetpulse.api@gmail.com
- * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
- */
-package swagger
-
-import (
-	"fmt"
-	"net/http"
-	"strings"
-
-	"github.com/gorilla/mux"
-)
-
-type Route struct {
-	Name        string
-	Method      string
-	Pattern     string
-	HandlerFunc http.HandlerFunc
-}
-
-type Routes []Route
-
-func NewRouter() *mux.Router {
-	router := mux.NewRouter().StrictSlash(true)
-	for _, route := range routes {
-		var handler http.Handler
-		handler = route.HandlerFunc
-		handler = Logger(handler, route.Name)
-
-		router.
-			Methods(route.Method).
-			Path(route.Pattern).
-			Name(route.Name).
-			Handler(handler)
-	}
-
-	return router
-}
-
-func Index(w http.ResponseWriter, r *http.Request) {
-	fmt.Fprintf(w, "Hello World!")
-}
-
-var routes = Routes{
-	Route{
-		"Index",
-		"GET",
-		"/v1/",
-		Index,
-	},
-
-	Route{
-		"Co2WeeklyGet",
-		strings.ToUpper("Get"),
-		"/v1/co2/weekly",
-		Co2WeeklyGet,
-	},
-
-	Route{
-		"Co2WeeklyIdGet",
-		strings.ToUpper("Get"),
-		"/v1/co2/weekly/{id}",
-		Co2WeeklyIdGet,
-	},
-
-	Route{
-		"HealthGet",
-		strings.ToUpper("Get"),
-		"/v1/health",
-		HealthGet,
-	},
-}
